+++ cmd/zetae2e/local/local.go
@@ -745,6 +745,9 @@ func waitKeygenHeight(
 		return errors.Wrap(err, "observerClient.Keygen error")
 	case resp.Keygen == nil:
 		return errors.New("keygen is nil")
+	case resp.Keygen.Status == observertypes.KeygenStatus_KeyGenSuccess:
+		// noop
+		return nil
 	case resp.Keygen.Status != observertypes.KeygenStatus_PendingKeygen:
 		return errors.Errorf("keygen is not pending (status: %s)", resp.Keygen.Status.String())
 	}
